Refactor to use contract interface
