+++ src/chains/types.ts
@@ -1,6 +1,12 @@
 import { type KeyPair } from '@near-js/crypto'
 import type BN from 'bn.js'
 
+/** 
+Available ChainSignature contracts:
+  - Mainnet: v1.signer
+  - Testnet: v1.signer-prod.testnet
+  - Development (unstable): v1.signer-dev.testnet
+*/
 export type ChainSignatureContracts = string
 
 export interface ChainProvider {
