+++ src/signature/index.ts
@@ -1,1 +1,1 @@
-export * from './signature'
+export * from './chain-signatures-contract'
