+++ src/signature/types.ts
@@ -3,3 +3,13 @@ export interface RSVSignature {
   s: string
   v: number
 }
+
+export interface MPCSignature {
+  big_r: {
+    affine_point: string
+  }
+  s: {
+    scalar: string
+  }
+  recovery_id: number
+}
